Added Redis session store and auth provider
This commit adds two Redis-backed things: an auth provider (handy for testing without LDAP) and a session store
so sessions can persist over  restarts. This also makes it easier to develop since you don't have to log in
all the time.

Additionally, a bootstrapper utility is provided to create an admin user in empty systems. It will also recreate
the admin user if needed.