Very rough first attempt at a React-based UI
The UI is alright, but it's ugly and could definitely do with being more dynamic.

In the spirit of learning new things and trying things out, I'm going to attempt
to write a React-based webapp in Typescript. I haven't used either of these
technologies before so wish me luck!

The idea here is to make the Go server even more of a pure API host, with the
client-side app doing all the rendering.