StatusWatcher: Ignore time when calculating health of transient node
Transient nodes come and go as the tool is used. This means that the node health cannot sensibly include time in its metrics, since this raises needless alerts when the tool is powered down and started up.

This commit makes considering last seen time conditional on the node not being transient.