Apply leniency already applied to doors when MQTT is disconnected
A fix was applied to doors some time ago to give them much more leniency on the API connection times.
This fix, however, only applied if MQTT was healthy. This commit applies the same fix to doors
when they are not seen on MQTT, and API times are the only thing we have to go on.

This was causing oscillations, since only when a door was used would it typically checkin with the API